Update the local dentry for deleted files before notifying the kernel.
This avoids a race condition where fstat() could fail on a recently
deleted file only.  To avoid this set the deleted flag before replying.
This does however mean there's a (potentially negative) hashtable
lookup as part of the unlink call.

Update the open_rename() and others tests now that it's expected to pass.

Signed-off-by: Ashley Pittman <ashley.m.pittman@intel.com>